mysite/requirements.txt
asgiref==3.2.7
Django==3.0.7
django-crispy-forms==1.9.1
django-taggit==1.3.0
Pillow==7.1.2
pytz==2019.3
sqlparse==0.3.1
